+++ FrontEnd/HRHJ/app/src/main/java/com/example/hrhj/Add/AddBasicFragment.java
@@ -1,6 +1,5 @@
-package com.example.hrhj;
+package com.example.hrhj.Add;
 
-import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
 
@@ -9,21 +8,17 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.GridView;
-import android.widget.ImageView;
 
+import com.example.hrhj.R;
 
-public class AddGalleryFragment extends Fragment {
+public class AddBasicFragment extends Fragment {
 
-    private ImageView galleryPreview;
-    private GridView gallery;
     private OnFragmentInteractionListener mListener;
 
-    public static AddGalleryFragment newInstance() {
-        return  new AddGalleryFragment();
+    public static AddBasicFragment newInstance() {
+        return new AddBasicFragment();
     }
-
-    public AddGalleryFragment() {
+    public AddBasicFragment() {
         // Required empty public constructor
     }
 
@@ -37,9 +32,7 @@ public class AddGalleryFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_add_gallery, container, false);
-
-        return view;
+        return inflater.inflate(R.layout.fragment_add_basic, container, false);
     }
 
     public interface OnFragmentInteractionListener {
