+++ FrontEnd/HRHJ/app/src/main/java/com/example/hrhj/Add/AddGalleryFragment.java
@@ -1,6 +1,5 @@
-package com.example.hrhj;
+package com.example.hrhj.Add;
 
-import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
 
@@ -9,15 +8,23 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.GridView;
+import android.widget.ImageView;
 
-public class AddBasicFragment extends Fragment {
+import com.example.hrhj.R;
 
+
+public class AddGalleryFragment extends Fragment {
+
+    private ImageView galleryPreview;
+    private GridView gallery;
     private OnFragmentInteractionListener mListener;
 
-    public static AddBasicFragment newInstance() {
-        return new AddBasicFragment();
+    public static AddGalleryFragment newInstance() {
+        return  new AddGalleryFragment();
     }
-    public AddBasicFragment() {
+
+    public AddGalleryFragment() {
         // Required empty public constructor
     }
 
@@ -31,7 +38,9 @@ public class AddBasicFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_add_basic, container, false);
+        View view = inflater.inflate(R.layout.fragment_add_gallery, container, false);
+
+        return view;
     }
 
     public interface OnFragmentInteractionListener {
