+++ FrontEnd/HRHJ/app/src/main/java/com/example/hrhj/Add/AddTabAdapter.java
@@ -1,14 +1,11 @@
-package com.example.hrhj;
+package com.example.hrhj.Add;
 
-import android.content.Context;
 import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentPagerAdapter;
-import androidx.fragment.app.FragmentStatePagerAdapter;
-import androidx.fragment.app.FragmentTransaction;
 
 public class AddTabAdapter extends FragmentPagerAdapter {
