+++ FrontEnd/HRHJ/app/src/main/java/com/example/hrhj/Add/CameraPreview.java
@@ -1,4 +1,4 @@
-package com.example.hrhj;
+package com.example.hrhj.Add;
 
 import android.content.Context;
 import android.content.Intent;
