+++ FrontEnd/HRHJ/app/src/main/res/layout/fragment_add.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@color/colorPrimary"
-    tools:context=".AddFragment">
+    tools:context=".Add.AddFragment">
 
 <!--    <TextView-->
 <!--        android:id="@+id/addDate"-->
