+++ FrontEnd/HRHJ/app/src/main/res/layout/fragment_add_camera.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddCameraFragment">
+    tools:context=".Add.AddCameraFragment">
 
     <SurfaceView
         android:id="@+id/surfaceView"
