+++ FrontEnd/HRHJ/app/src/main/res/layout/fragment_add_gallery.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".AddGalleryFragment">
+    tools:context=".Add.AddGalleryFragment">
 
     <ImageView
         android:id="@+id/galleryPreview"
